+++ supermage/utils/math_utils.py
@@ -1,7 +1,7 @@
 import torch
 
-def _func_R_x(theta, device = "cuda"):
-    result = torch.zeros((3, 3), device = device, dtype = torch.float64)
+def _func_R_x(theta, device = "cuda", dtype = torch.float64):
+    result = torch.zeros((3, 3), device = device, dtype = dtype)
     result[0, 0] = 1
     result[1, 1] = torch.cos(theta)
     result[1, 2] = -torch.sin(theta)
@@ -9,8 +9,8 @@ def _func_R_x(theta, device = "cuda"):
     result[2, 2] = torch.cos(theta)
     return result
 
-def _func_R_y(psi, device = "cuda"):
-    result = torch.zeros((3, 3), device = device, dtype = torch.float64)
+def _func_R_y(psi, device = "cuda", dtype = torch.float64):
+    result = torch.zeros((3, 3), device = device, dtype = dtype)
     result[0, 0] = torch.cos(psi)
     result[0, 2] = torch.sin(psi)
     result[1, 1] = 1
@@ -18,8 +18,8 @@ def _func_R_y(psi, device = "cuda"):
     result[2, 2] = torch.cos(psi)
     return result
 
-def _func_R_z(phi, device = "cuda"):
-    result = torch.zeros((3, 3), device = device, dtype = torch.float64)
+def _func_R_z(phi, device = "cuda", dtype = torch.float64):
+    result = torch.zeros((3, 3), device = device, dtype = dtype)
     result[0, 0] = torch.cos(phi)
     result[0, 1] = -torch.sin(phi)
     result[1, 0] = torch.sin(phi)
@@ -27,11 +27,11 @@ def _func_R_z(phi, device = "cuda"):
     result[2, 2] = 1
     return result
 
-def DoRotation(x, y, z, theta, phi, device = "cuda"):
+def DoRotation(x, y, z, theta, phi, device = "cuda", dtype = torch.float64):
     """Rotate a meshgrid. Theta is inclination, phi is rotation in sky plane"""
     shape = x.shape
     # Clockwise, 2D rotation matrix
-    RotMatrix = torch.matmul(_func_R_y(theta, device = device), _func_R_z(phi, device = device))
+    RotMatrix = torch.matmul(_func_R_y(theta, device = device, dtype = dtype), _func_R_z(phi, device = device, dtype = dtype))
     #return torch.einsum('ji, mni -> jmn', RotMatrix, torch.dstack([x, y, z]))
     
     mult = torch.inner(RotMatrix, torch.dstack([x.ravel(),y.ravel(), z.ravel()]))
@@ -41,11 +41,11 @@ def DoRotation(x, y, z, theta, phi, device = "cuda"):
     return xrot, yrot, zrot
 
 
-def DoRotationT(x, y, z, theta, phi, device = "cuda"):
+def DoRotationT(x, y, z, theta, phi, device = "cuda", dtype = torch.float64):
     """Multiply a meshgrid by the transpose of rotation matrix. Theta is inclination, phi is rotation in sky plane"""
     shape = x.shape
     # Clockwise, 2D rotation matrix
-    RotMatrix = torch.matmul(_func_R_y(theta, device = device), _func_R_z(phi, device = device)).T
+    RotMatrix = torch.matmul(_func_R_y(theta, device = device), _func_R_z(phi, device = device, dtype = dtype)).T
     #return torch.einsum('ji, mni -> jmn', RotMatrix, torch.dstack([x, y, z]))
     mult = torch.inner(RotMatrix, torch.dstack([x.ravel(),y.ravel(), z.ravel()]))
     xrot = mult[0,:].reshape(shape)
